requirements.txt
pandas==2.2.3
openpyxl==3.1.5
matplotlib==3.10.0
requests==2.32.3
yfinance==0.2.50
python-dotenv==1.1.0
